tabletserver endtoend: transaction tests WIP
