+++ go/vt/tabletserver/endtoend/framework/logtracker.go
@@ -1,22 +1,37 @@
-// Copyright 2012, Google Inc. All rights reserved.
+// Copyright 2015, Google Inc. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
 package framework
 
-import "github.com/youtube/vitess/go/vt/tabletserver"
+import (
+	"time"
+
+	"github.com/youtube/vitess/go/vt/tabletserver"
+)
 
 // TxFetcher alows you to capture and fetch info on the next transaction.
-type TxFetcher chan interface{}
+type TxFetcher struct {
+	start time.Time
+	ch    chan interface{}
+}
 
 // NewTxFetcher sets up the capture and returns a new TxFetcher.
-func NewTxFetcher() TxFetcher {
-	return TxFetcher(tabletserver.TxLogger.Subscribe("endtoend"))
+func NewTxFetcher() *TxFetcher {
+	return &TxFetcher{
+		start: time.Now(),
+		ch:    tabletserver.TxLogger.Subscribe("endtoend"),
+	}
 }
 
 // Fetch fetches the last captured transaction. It must be called only once.
-func (ch TxFetcher) Fetch() *tabletserver.TxConnection {
-	tx := (<-ch).(*tabletserver.TxConnection)
-	tabletserver.TxLogger.Unsubscribe(ch)
-	return tx
+func (fetcher *TxFetcher) Fetch() *tabletserver.TxConnection {
+	for {
+		tx := (<-fetcher.ch).(*tabletserver.TxConnection)
+		// Skip any events that pre-date start time.
+		if tx.EndTime.After(fetcher.start) {
+			tabletserver.TxLogger.Unsubscribe(fetcher.ch)
+			return tx
+		}
+	}
 }
