+++ go/vt/tabletserver/endtoend/nocache_test.go
@@ -6,6 +6,7 @@ package endtoend
 
 import (
 	"reflect"
+	"strings"
 	"testing"
 
 	"github.com/youtube/vitess/go/mysql"
@@ -152,99 +153,18 @@ func TestNocacheListArgs(t *testing.T) {
 	}
 }
 
-func TestCommit(t *testing.T) {
+func TestIntegrityError(t *testing.T) {
 	client := framework.NewDefaultClient()
-	defer client.Execute("delete from vtocc_test where intval=4", nil)
-
-	fetcher := framework.NewTxFetcher()
 	vstart := framework.DebugVars()
-
-	query := "insert into vtocc_test (intval, floatval, charval, binval) " +
-		"values(4, null, null, null)"
-	err := client.Begin()
-	if err != nil {
-		t.Error(err)
-		return
+	_, err := client.Execute("insert into vtocc_test values(1, null, null, null)", nil)
+	want := "error: Duplicate entry '1'"
+	if err == nil || !strings.HasPrefix(err.Error(), want) {
+		t.Errorf("Error: %v, want prefix %s", err, want)
 	}
-	_, err = client.Execute(query, nil)
-	if err != nil {
-		t.Error(err)
-		return
-	}
-	err = client.Commit()
-	if err != nil {
-		t.Error(err)
-		return
-	}
-	tx := fetcher.Fetch()
-	want := []string{query}
-	if !reflect.DeepEqual(tx.Queries, want) {
-		t.Errorf("queries: %v, want %v", tx.Queries, want)
-	}
-
-	qr, err := client.Execute("select * from vtocc_test", nil)
-	if err != nil {
-		t.Error(err)
-		return
-	}
-	if qr.RowsAffected != 4 {
-		t.Errorf("rows affected: %d, want 4", qr.RowsAffected)
-	}
-
-	_, err = client.Execute("delete from vtocc_test where intval=4", nil)
-	if err != nil {
-		t.Error(err)
-		return
-	}
-
-	qr, err = client.Execute("select * from vtocc_test", nil)
-	if err != nil {
-		t.Error(err)
-		return
-	}
-	if qr.RowsAffected != 3 {
-		t.Errorf("rows affected: %d, want 4", qr.RowsAffected)
-	}
-
 	vend := framework.DebugVars()
-	v1 := framework.FetchInt(vstart, "Transactions.TotalCount")
-	v2 := framework.FetchInt(vend, "Transactions.TotalCount")
-	if v1+2 != v2 {
-		t.Errorf("Transactions.TotalCount: %d, want %d", v2, v1+2)
-	}
-	v1 = framework.FetchInt(vstart, "Transactions.Histograms.Completed.Count")
-	v2 = framework.FetchInt(vend, "Transactions.Histograms.Completed.Count")
-	if v1+2 != v2 {
-		t.Errorf("Transactions.Histograms.Completed.Count: %d, want %d", v2, v1+2)
-	}
-	v1 = framework.FetchInt(vstart, "Queries.TotalCount")
-	v2 = framework.FetchInt(vend, "Queries.TotalCount")
-	if v1+6 != v2 {
-		t.Errorf("Queries.TotalCount: %d, want %d", v2, v1+6)
-	}
-	v1 = framework.FetchInt(vstart, "Queries.Histograms.BEGIN.Count")
-	v2 = framework.FetchInt(vend, "Queries.Histograms.BEGIN.Count")
+	v1 := framework.FetchInt(vstart, "InfoErrors.DupKey")
+	v2 := framework.FetchInt(vend, "InfoErrors.DupKey")
 	if v1+1 != v2 {
-		t.Errorf("Queries.Histograms.BEGIN.Count: %d, want %d", v2, v1+1)
-	}
-	v1 = framework.FetchInt(vstart, "Queries.Histograms.COMMIT.Count")
-	v2 = framework.FetchInt(vend, "Queries.Histograms.COMMIT.Count")
-	if v1+1 != v2 {
-		t.Errorf("Queries.Histograms.COMMIT.Count: %d, want %d", v2, v1+1)
-	}
-	v1 = framework.FetchInt(vstart, "Queries.Histograms.INSERT_PK.Count")
-	v2 = framework.FetchInt(vend, "Queries.Histograms.INSERT_PK.Count")
-	if v1+1 != v2 {
-		t.Errorf("Queries.Histograms.INSERT_PK.Count: %d, want %d", v2, v1+1)
-	}
-	v1 = framework.FetchInt(vstart, "Queries.Histograms.DML_PK.Count")
-	v2 = framework.FetchInt(vend, "Queries.Histograms.DML_PK.Count")
-	if v1+1 != v2 {
-		t.Errorf("Queries.Histograms.DML_PK.Count: %d, want %d", v2, v1+1)
-	}
-	v1 = framework.FetchInt(vstart, "Queries.Histograms.PASS_SELECT.Count")
-	v2 = framework.FetchInt(vend, "Queries.Histograms.PASS_SELECT.Count")
-	if v1+2 != v2 {
-		t.Errorf("Queries.Histograms.PASS_SELECT.Count: %d, want %d", v2, v1+2)
+		t.Errorf("InfoErrors.DupKey: %d, want %d", v2, v1+1)
 	}
 }
